backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# WebSockets
websockets==13.1

# Firestore
google-cloud-firestore==2.19.0

# Firebase Auth (for SDK, even though frontend handles auth)
firebase-admin==6.6.0

# Pydantic
pydantic==2.10.3
pydantic-settings==2.7.0

# Environment
python-dotenv==1.0.1